# Issue152 vectorize depth pipe methods