Closes #152.