@MassimoCimmino 2021-11-16
## Timing results using `SingleUTube` and `MultipleUTube` (`nPipes=2, config='series'`) classes.
```
>>> m_flow = 0.05
>>> m_flow_random = (np.random.rand(100) + 1.) * m_flow
>>> z = np.linspace(0., 1., num=50) * SingleUTube.b.H
>>> T_in = 5.
>>> T_b_12 = 10. * np.ones(12)
>>> T_b_1 = 10.
>>> %timeit for m_flow_i in m_flow_random: SingleUTube.coefficients_borehole_heat_extraction_rate(m_flow_i, 4000., nSegments=12)
>>> %timeit for m_flow_i in m_flow_random: SingleUTube.coefficients_borehole_heat_extraction_rate(m_flow_i, 4000., nSegments=1)
>>> %timeit for m_flow_i in m_flow_random: SingleUTube.get_temperature(z, T_in, T_b_12, m_flow_i, 4000.)
>>> %timeit for m_flow_i in m_flow_random: SingleUTube.get_temperature(z, T_in, T_b_1, m_flow_i, 4000.)
>>> %timeit for m_flow_i in m_flow_random: SingleUTube.coefficients_borehole_heat_extraction_rate(m_flow_i, 4000., nSegments=1)
>>> %timeit for m_flow_i in m_flow_random: DoubleUTube_ser.coefficients_borehole_heat_extraction_rate(m_flow_i, 4000., nSegments=12)
>>> %timeit for m_flow_i in m_flow_random: DoubleUTube_ser.coefficients_borehole_heat_extraction_rate(m_flow_i, 4000., nSegments=1)
>>> %timeit for m_flow_i in m_flow_random: DoubleUTube_ser.get_temperature(z, T_in, T_b_12, m_flow_i, 4000.)
>>> %timeit for m_flow_i in m_flow_random: DoubleUTube_ser.get_temperature(z, T_in, T_b_1, m_flow_i, 4000.)
```

## Before changes
```
212 ms ± 1.99 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
52.8 ms ± 185 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
721 ms ± 3.89 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
698 ms ± 2.43 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
51.7 ms ± 218 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
170 ms ± 985 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
69.2 ms ± 521 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
482 ms ± 5.04 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
458 ms ± 2.85 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

## After changes
```
45.3 ms ± 328 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
42.1 ms ± 232 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
43.8 ms ± 399 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
38.6 ms ± 329 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
42.7 ms ± 423 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
66.3 ms ± 763 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
63.2 ms ± 117 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
66.6 ms ± 590 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
60.4 ms ± 581 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```